+++ CHANGELOG.md
@@ -21,6 +21,8 @@ This project adheres to Semantic Versioning.
 * `Vec2` is now exported from `math`, not `graphics`.
 * The graphics device debugging info is now hidden by default. Set the `debug_info` option on `Game` to `true` to bring this back.
 * Details of the active graphics device can now be retrieved by calling `graphics::get_device_info`.
+* The functions for setting the fullscreen/cursor visibility state have been changed to take booleans, instead of there being multiple functions.
+* Functions for getting and setting vsync have been added to `window`.
 * Updated `glow` to 0.2.3.
 * Updated `hashbrown` to 0.6.1.
 * Updated `image` to 0.22.3.